chore(*): remove codeBlockSource
we wish for code example to be better. however, passing source as string
is no better than having it in readme.md or as part of `examples`.
moreover, it loses interactivity, playing with props no longer reflects
in code.

code example should be better but with a different approach. thus,
removing `codeBlockSource` for now