<LiveCodeExample/> - add prettier
This PR runs prettier in the browser on the `initialCode` being passed to the `<LiveCodeExample/>`.

*Why?*

The `initialCode` being passed may be built dynamically. For example, consider the following example in WSR:

https://github.com/wix/wix-style-react/blob/master/stories/Popover/examples/ExampleAppendTo.js#L11-L48

A utility function called `createPropsArray` is used to create a string representing the props, and the result may not follow out coding style. A way to solve this is to add `prettier` to format that initial code.

Also, why not? 🙃