fix testkit parser failure
this is a hack to prevent testkit parser to crash, when unknown node
type is reached.

unknown node type can be reached due to numerous reasons and
implementing all of them is much more costly than allowing to silently
fail.

silent failure in this case is okay, because we can still get useful
information about not failed items, whereas if parser fails entirely, we
get no information at all